+++ disassemblers/ofrak_ghidra/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## 0.1.1 - 2024-02-15
 ### Added
 - Added typing support to the ofrak-ghidra package. This is helpful for users who use `mypy` and `ofrak_ghidra` in a project.
+- Added GhidraDecompilationAnalyzer to retrieve the decompilation of a complex block. ([#453](https://github.com/redballoonsecurity/ofrak/pull/453))
 
 ## 0.1.0 - 2022-08-09
 ### Added